+++ contracts/Escrow.sol
@@ -470,8 +470,7 @@ contract Escrow is InitializableERC20 {
         uint48 earliestExerciseTime = SafeCast.toUint48(
             block.timestamp + auctionParams.earliestExerciseTenor
         );
-
-        (uint256 matchFeeProtocol, uint256 matchFeeDistPartner) = Router(router)
+        (uint128 matchFeeProtocol, uint128 matchFeeDistPartner) = Router(router)
             .getMatchFees(distPartner, premium);
 
         if (matchFeeProtocol + matchFeeDistPartner >= premium) {
